♻️🎪 Split application page
