+++ apps/show/src/routes/_layout.exposants.candidature._index/loader.server.ts
@@ -1,0 +1,10 @@
+import { notFound } from "#core/response.server";
+import { json } from "@remix-run/node";
+
+export async function loader() {
+  if (process.env.FEATURE_FLAG_EXHIBITOR_APPLICATION_ONLINE !== "true") {
+    throw notFound();
+  }
+
+  return json("ok" as const);
+}
