+++ apps/show/src/routes/_layout.exposants.candidature._index/route.tsx
@@ -1,22 +1,14 @@
 import { ErrorPage, getErrorTitle } from "#core/data-display/error-page";
 import { createSocialMeta } from "#core/meta";
 import { getPageTitle } from "#core/page-title";
-import { notFound } from "#core/response.server";
-import { json } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
+import type { loader } from "./loader.server";
 import { SectionDescription } from "./section-description";
 import { SectionForm } from "./section-form";
 import { SectionTitle } from "./section-title";
 
 export { action } from "./action.server";
-
-export async function loader() {
-  if (process.env.FEATURE_FLAG_EXHIBITOR_APPLICATION_ONLINE !== "true") {
-    throw notFound();
-  }
-
-  return json("ok" as const);
-}
+export { loader } from "./loader.server";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return createSocialMeta({
